+++ mountpoint-s3-fs/CHANGELOG.md
@@ -1,4 +1,7 @@
-## Unreleased
+## Unreleased (v0.8.4)
+
+* Add metric to track cache hit rate in logs. ([#1716](https://github.com/awslabs/mountpoint-s3/pull/1716))
+* Remove redundant cache merics in logs. ([#1716](https://github.com/awslabs/mountpoint-s3/pull/1716))
 
 ## v0.8.3 (October 30, 2025)